feat: README's for all langs, dynamic generation of flake Large change!
First off mkTemplates will now conveniently create the proper templates
for the flake, so all I need to do is specify the path (which is
assumed to be the name of the language too).

Secondly I added READMEs to all language templates, but made them get
filtered out by the flake. This is so that I can put helpful information
in the READMEs without it being propagated to the flake itself.